+++ model/jobresult.go
